+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). 
 - Black ops and Theft success rates now use the same formula (~ average of the two)
 - Adjust XP gain for invasion to 1 per acre gained (from 1 per acre captured)
 - Adjust XP gain for info/black/war to 2/4/6 (from 1/3/5)
+- Heroes now lose 1 XP per acre lost from invasion
 - Human Cavalry: +50p (from 1200p)
 
 ## [1.32.0] - 2022-11-01